Put log function back into logger class.
