+++ training/src/anemoi/training/diagnostics/mlflow/logger.py
@@ -23,17 +23,22 @@ from typing import Any
 from typing import Literal
 from weakref import WeakValueDictionary
 
+from packaging.version import Version
 from pytorch_lightning.loggers.mlflow import MLFlowLogger
+from pytorch_lightning.loggers.mlflow import _convert_params
+from pytorch_lightning.loggers.mlflow import _flatten_dict
 from pytorch_lightning.utilities.rank_zero import rank_zero_only
 
 from anemoi.training.diagnostics.mlflow.auth import TokenAuth
+from anemoi.training.diagnostics.mlflow.utils import clean_config_params
+from anemoi.training.diagnostics.mlflow.utils import expand_iterables
 from anemoi.training.diagnostics.mlflow.utils import health_check
-from anemoi.training.diagnostics.mlflow.utils import log_hyperparams_in_mlflow
 
 if TYPE_CHECKING:
     from argparse import Namespace
 
     import mlflow
+    from mlflow.tracking import MlflowClient
 
 LOGGER = logging.getLogger(__name__)
 
@@ -510,7 +515,7 @@ class AnemoiMLflowLogger(MLFlowLogger):
             have lists converted according to `expand_iterables`,
             by default None.
         """
-        log_hyperparams_in_mlflow(
+        AnemoiMLflowLogger.log_hyperparams_in_mlflow(
             self.experiment,
             self.run_id,
             params,
@@ -530,3 +535,94 @@ class AnemoiMLflowLogger(MLFlowLogger):
             self.run_id_to_log_monitor[self.run_id].finish(status)
 
         super().finalize(status)
+
+    @staticmethod
+    def log_hyperparams_in_mlflow(
+        client: MlflowClient,
+        run_id: str,
+        params: dict[str, Any] | Namespace,
+        *,
+        expand_keys: list[str] | None = None,
+        log_hyperparams: bool | None = True,
+        clean_params: bool = True,
+    ) -> None:
+        """Log hyperparameters to MLflow server.
+
+        - flatten config params using '.'.
+        - expand keys within params to avoid truncation.
+        - log hyperparameters as an artifact.
+
+        Parameters
+        ----------
+        client : MlflowClient
+            MLflow client.
+        run_id : str
+            Run ID.
+        params : dict[str, Any] | Namespace
+            params to log.
+        expand_keys : list[str] | None, optional
+            keys to expand within params. Any key being expanded will
+            have lists converted according to `expand_iterables`,
+            by default None.
+        log_hyperparams : bool | None, optional
+            Whether to log hyperparameters, by default True.
+        """
+        if log_hyperparams:
+            params = _convert_params(params)
+
+            # this is needed to resolve optional missing config values to a string, instead of raising a missing error
+            if config := params.get("config"):
+                params["config"] = config.model_dump(by_alias=True)
+
+            import mlflow
+            from mlflow.entities import Param
+
+            truncation_length = 250
+
+            if Version(mlflow.VERSION) >= Version("1.28.0"):
+                truncation_length = 500
+
+            AnemoiMLflowLogger.log_hyperparams_as_mlflow_artifact(client=client, run_id=run_id, params=params)
+
+            expanded_params = {}
+            params = params.copy()
+
+            for key in expand_keys or []:
+                if key in params:
+                    expanded_params.update(
+                        expand_iterables(params.pop(key), size_threshold=None, delimiter="."),
+                    )
+            expanded_params.update(params)
+
+            expanded_params = _flatten_dict(
+                expanded_params,
+                delimiter=".",
+            )  # Flatten dict with '.' to not break API queries
+            if clean_params:
+                expanded_params = clean_config_params(expanded_params)
+
+            # Truncate parameter values.
+            params_list = [Param(key=k, value=str(v)[:truncation_length]) for k, v in expanded_params.items()]
+            for idx in range(0, len(params_list), 100):
+                client.log_batch(run_id=run_id, params=params_list[idx : idx + 100])
+
+    @staticmethod
+    def log_hyperparams_as_mlflow_artifact(
+        client: MlflowClient,
+        run_id: str,
+        params: dict[str, Any] | Namespace,
+    ) -> None:
+        """Log hyperparameters as an artifact."""
+        import json
+        import tempfile
+        from json import JSONEncoder
+
+        class StrEncoder(JSONEncoder):
+            def default(self, o: Any) -> str:
+                return str(o)
+
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            path = Path(tmp_dir) / "config.json"
+            with Path.open(path, "w") as f:
+                json.dump(params, f, cls=StrEncoder)
+            client.log_artifact(run_id=run_id, local_path=path)
